split code for lib
